inspectors: validate artefacts using opinions
Instead if having a certain format being examined only by one
inspector which would then ask other inspectors to validate the
artefact, let inspectors examine all related formats and emit
opinions. An artefact is validated only if approved by all
inspectors that can opine. For example, the inspection of a
deb package downloaded before updating apt metadata could
result in the following set of opinions:
```
  "opinions": [
           {
                   "inspector-id": "deb",
                   "opinion": "Approved",
                   "reason": "deb package parsed"
           },
           {
                   "inspector-id": "apt.packages",
                   "opinion": "Rejected",
                   "reason": "deb file digest not listed in packages file"
           }
   ],
```
This means that the deb package itself can be consistent, but
the apt inspector rejected it because its digest was not found
in the apt metadata.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>