service: add proxy acl support (#90)
Proxy ACLs are specified as a separate configuration file, reloaded
by the fetch service when modified. Connections to servers are filtered
by rules specified in the configuration file: each rule contains a list
of IP addresses or networks, and an access allow/deny entry. The first
match of the destination address tells if the connection will be allowed
of denied.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>