feat: add Maven POM inspector
POM-based projects declared their dependencies by version; as such, when
building the project Maven accesses the POM for each dependency through
a fully-qualified URL that includes the group, artifact, and version.

This means that this inspector can be very simple because there is no
"index" that Maven queries to determine versions; instead, we just
inspect the downloaded POMs.

Fixes #168