fix(inspectors/git): handle absence of proto and command
If the git protocol message is missing, the request inspector won't
be able to set the command annotation. The artefact inspectors must not
error in case of command not being set, and just ignore the artefact in
this case.

Additionally, reject the artefact instead of generating a fatal error in
case of git objects unpacking error, and refuse to process unshallow
requests.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>