refactor: only unpack git repos once
With this change, the git.upload-pack inspector becomes in charge of doing all
git-related verifications (whether the fetch is shallow, single ref only, etc)
and unpacks & checks out the wanted ref into a session-local directory that
persists after the inspector is done. Other git-related inspectors can then use
the new 'git-checkout-pack' annotation to directly inspect the contents of the
repository, without duplicating the work.

This requires the introduction of a new "CacheDir" field in artefacts, so that
the inspector has a place to put files that a) persists after that inspector
is done and b) gets removed when the session's resources are cleaned-up.

Fixes #276