fix(inspectors/deb): remove incorrect metadata mapping
We can't assume that the name marked as the `Upstream-Contact` in the
machine-readable copyright file is necessarily the upstream author, so
remove this mapping. The same applies to the mapping from maintainer
to author email. Additionally, these two pieces of information come
from different origins and may be inconsistent, with a name listed as
author with a completely different address listed in author-email.

As a general rule we need to be less aggressive in assuming things
when extracting information during inspection, to avoid false positives.

Fixes #324.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>