feat: use different defaults when in snap
When running as a snap, use ${SNAP_DATA}/conf as the default config dir and
${SNAP_COMMON}/spool as the default spool dir. This is necessary because the
non-snap locations (/etc/fetch and /var/lib/fetch) are not accessible when
running as a strict snap.

This change entails some minor refactoring to make this part of the code more
testable, without the "singleton" opts object.