feat!: handle "default" origins as a special case
We currently have many inspectors that are not configurable and accept their
respective ecosystem's default "registries" (crates.io, pypi, etc). This is
not acceptable in a strict session where all sources my be explicitly declared.

Since there are currently consumers of the session report that would be
affected if we just disallowed default registries, change these instances to
"Unknown" request inspections so that permissive sessions still get these
artifacts. This is done through a new helper function that can be applied to
"currently unconfigurable" inspectors and then removed once configuration is
implemented.

This is a breaking change because strict sessions will no longer accept these
'default' registries.