fix: do not register the chisel inspector yet
The chisel.release inspector needs to have a configurable release URL and some
validation logic to check if archives defined in the chisel.yaml file are OK.

The plan is to add that logic in a follow-up PR and merge this PR right now.
However, since this PR is incomplete, this commit unregisters the chisel
inspector from the fetch service.