feat: add OCI image inspector
This commit adds a new OCI image inspector which checks requests to and
from OCI registries such as Dockerhub, ECR etc. It relies on an ideal
OCI image registry as described by the OCI distribution spec[^1].

It expects the following configuration:

    oci:
      registries:
	<name>:
	  url: <registry-url>
	  auth-url: <registry-authentication-url>

With an ideal registry, a client should make the following network
requests which the OCI inspector checks:

  1. Ping the registry at the /v2/ endpoint to see if the registry
     supports the OCI distribution spec.[^1] This typically results in a
     401 Unauthorized response with the "www-authenticate" header set to
     the authentication URL. Since inspectors do not get to see the
     response headers on a 401 response, the OCI inspector takes in an
     "auth-url" configuration parameter which describes the
     authentication URL, as "www-authenticate" would set it.

  2. Make a request to the authentication URL to generate access token.

  3. Pull the image index file using the following endpoint:
     ...    /v2/<image-name>/manifests/<reference>     ("end-3" at [^1])
     <reference> can either be a tag or digest. The client would then
     parse the image index to get the digests for the manifests.

  4. Pull the image manifest using the above ("end-3") endpoint. The
     client would then parse the manifest file to get the digests of the
     config file and the layers blob.

  5. Pull the config and layer blobs, using the following endpoint:
     ...    /v2/<image-name>/blobs/<digest>            ("end-2" at [^1])

If an OCI registry returns a 307 response with "Location" header to
request 5 ("end-2"), as Dockerhub and ECR typically do, the OCI
inspector needs to be run in "permissive mode" to approve those requests
since the OCI inspector is stateful (like apt inspectors) and keeps info
about which requests to approve.

[^1] https://github.com/opencontainers/distribution-spec/blob/main/spec.md#endpoints