fix: address download timeout issues
When a response is processed, the Fetch Service downloads the entire
body contents to a local buffer, inspects it, and then grants the
original client access to the downloaded data. Depending on the volume
of the data being transferred and link conditions, it may take some time
for the client to get a response. This may cause timeouts and prevent
files to be downloaded.

Clients typically time out quickly before receiving the response header,
and can wait more to actually receive body data. Snap downloads will
time out in 10s if a response header is not received. To address this
issue, responses processed in less than 5s will be immediately delivered
to the original client. If the download and inspection takes more than
that, we'll send a 200 response to the client and keep downloading data
in background. After the download finishes the data is inspected and
piped to the original client if it passes inspection. If inspection
fails, we close the connection. It's not polite, but the client will
know that something wrong happened when they get an unexpected EOF.

Downloads taking less than 5s will receive 403 in case of inspection
failure in strict mode.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>