fix(i/apt): prevent false positives in deb inspection (#422)
Deb artifacts are inspected by both the deb inspector (for the content)
and Packages inspector (to verify if the digest is listed in the
Packages file). Only the deb inspector should give the final approval,
with transactional inspectors rejecting if their verification fail.
This avoids the situation where the Packages inspector approves and
the main inspector ignores the artifact, causing an non-inspected
artifact to be downloaded by the original client. Also the deb inspector
requires the artifact to be previously inspected by the Packages
inspector.

The same happens between the Release inspector and artifacts listed
in the InRelease file (Packages, Commands, and Translation files).

Additionally, check if the Packages file is listed in the Release file
when verifying the deb file digest. This avoids the situation where a
deb file is correctly listed in an invalid Packages file.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>