feat: add inspector for the store resolve API (#450)
Implement a new inspector for the store resolve revisions API. The
existing info API was previously referred to as "api", rename that to
info API to avoid ambiguities.

The resolve API inspector is stateless. The purpose of the resolve
API is informative and not a security element pointing to temporary
URLs for authorization control, so there's no need to store resolved
data to check against subsequent requests.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>