feat: introduce apt repository aliases
The apt inspector is stateful and checks downloaded artifacts in
a chain of validations based on the repository URL. If the same
repository is accessed from two different URLs files are not
cross-validated. Additionally, the same file

This becomes a problem when the security repository is configured
to be accessed by default from security.ubuntu.com by apt and
archives.ubuntu.com by chisel: if an artifact such as the InRelease
file was already downloaded, it won't be downloaded again (and
re-inspected inside the new chain of validations for a different
URL).

In this case, we'll consider the two URLs as aliases for the same
repository. This is now expressed in the configuration entry for
that apt repository with the `BaseUrlAlias` key. Alias values
replace the host part of the artifact's URL, and will be used when
processing the validation chain.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>