fix: check session exists before completing inspection (#489)
A session can be finished before an inspection fully complete.
This can happen when a session timeouts at the same moment
an inspection completes.

Make sure the session is still active before formally completing
the inspection. If the session is not valid, the inspected artifact
and the result of the inspection are ignored.

Fixes #486

Signed-off-by: Paul Mars <paul.mars@canonical.com>